% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_post.R
\name{ci_multipliers}
\alias{ci_multipliers}
\title{Compute Std.err ==> conf.int multipliers.}
\usage{
ci_multipliers(lambda, sig2 = NULL, level = 0.95)
}
\arguments{
\item{lambda}{mean}

\item{sig2}{overdispersion parameter}

\item{level}{the confidence level required}
}
\value{
matrix with multipliers. col1=lo; col2=hi
}
\description{
Compute Std.err ==> conf.int multipliers.
}
