% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_outfile.R
\name{get_time_indices}
\alias{get_time_indices}
\title{Extract time indices from \code{tof} object}
\usage{
get_time_indices(x)
}
\arguments{
\item{x}{An object of class \code{tof}}
}
\value{
A \code{data.frame}
}
\description{
Extract time indices from \code{tof} object
}
\seealso{
Other parse_output: \code{\link{get_coef}},
  \code{\link{get_estimation_method}},
  \code{\link{get_gof}}, \code{\link{get_model}},
  \code{\link{get_n_site}}, \code{\link{get_n_time}},
  \code{\link{get_overal_imputed_slope}},
  \code{\link{get_time_totals}}, \code{\link{get_version}},
  \code{\link{get_wald}}, \code{\link{read_tof}},
  \code{\link{read_vcv}}
}
\concept{parse_output}
\keyword{internal}
