% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_outfile.R
\name{get_version}
\alias{get_version}
\title{Extract TRIM version used for output}
\usage{
get_version(x)
}
\value{
\code{character}
}
\description{
Extract TRIM version used for output
}
\seealso{
Other parse_output: \code{\link{get_coef}},
  \code{\link{get_estimation_method}},
  \code{\link{get_gof}}, \code{\link{get_model}},
  \code{\link{get_n_site}}, \code{\link{get_n_time}},
  \code{\link{get_overal_imputed_slope}},
  \code{\link{get_time_indices}},
  \code{\link{get_time_totals}}, \code{\link{get_wald}},
  \code{\link{read_tof}}, \code{\link{read_vcv}}
}
\concept{parse_output}
\keyword{internal}
