% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_post.R
\name{now_what}
\alias{now_what}
\title{Give advice on further refinement of TRIM models}
\usage{
now_what(z)
}
\arguments{
\item{z}{an object of class \code{\link{trim}}.}
}
\description{
Give advice on further refinement of TRIM models
}
\examples{

data(skylark)
z <- trim(count ~ site + time, data=skylark, model=2, overdisp=TRUE)
now_what(z)
}
\seealso{
\code{\link{trim}}

Other analyses: \code{\link{coef.trim}},
  \code{\link{confint.trim}}, \code{\link{gof}},
  \code{\link{index}}, \code{\link{overall}},
  \code{\link{overdispersion}},
  \code{\link{plot.trim.index}},
  \code{\link{plot.trim.overall}}, \code{\link{results}},
  \code{\link{serial_correlation}},
  \code{\link{summary.trim}}, \code{\link{totals}},
  \code{\link{trim}}, \code{\link{vcov.trim}},
  \code{\link{wald}}
}
