% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_post.R
\name{overdispersion}
\alias{overdispersion}
\title{Extract overdispersion from trim object}
\usage{
overdispersion(x)
}
\arguments{
\item{x}{An object of class \code{\link{trim}}}
}
\value{
The overdispersion value if computed, otherwise \code{NULL}.
}
\description{
Extract overdispersion from trim object
}
\seealso{
Other analyses: 
\code{\link{coef.trim}()},
\code{\link{confint.trim}()},
\code{\link{gof}()},
\code{\link{index}()},
\code{\link{now_what}()},
\code{\link{overall}()},
\code{\link{plot.trim.index}()},
\code{\link{plot.trim.overall}()},
\code{\link{plot.trim.smooth}()},
\code{\link{results}()},
\code{\link{serial_correlation}()},
\code{\link{summary.trim}()},
\code{\link{totals}()},
\code{\link{trendlines}()},
\code{\link{trim}()},
\code{\link{vcov.trim}()},
\code{\link{wald}()}
}
\concept{analyses}
