% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oystercatcher.R
\docType{data}
\name{oystercatcher}
\alias{oystercatcher}
\title{Oystercatcher population data}
\format{\code{.RData}}
\usage{
data(oystercatcher)
}
\description{
A sample data set for demonstation of monthly counts.

The \bold{\code{oystercatcher}} data set looks as follows.

\tabular{lll}{
 \bold{Column}      \tab \bold{Type}    \tab \bold{Description}\cr
 \code{site}        \tab \code{integer} \tab Site number\cr
 \code{year}        \tab \code{integer} \tab Year\cr
 \code{month}       \tab \code{integer} \tab Month\cr
 \code{count}       \tab \code{integer} \tab Counted oystercatchers
}
}
