% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tdf.R
\name{print.count.summary}
\alias{print.count.summary}
\title{print a count summary}
\usage{
\method{print}{count.summary}(x, ...)
}
\arguments{
\item{x}{An R object}

\item{...}{unused}
}
\description{
print a count summary
}
\keyword{internal}
