% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_gof.R
\name{print.trim.gof}
\alias{print.trim.gof}
\title{Print method for \code{trim.gof}}
\usage{
\method{print}{trim.gof}(x, ...)
}
\arguments{
\item{x}{a \code{trim.gof} object}
}
\description{
Print method for \code{trim.gof}
}
\keyword{internal}
