% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_overall.R
\name{print.trim.overall}
\alias{print.trim.overall}
\title{Print an object of class trim.overall}
\usage{
\method{print}{trim.overall}(x, ...)
}
\arguments{
\item{x}{An object of class \code{trim.overall}}
}
\description{
Print an object of class trim.overall
}
\keyword{internal}
