% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_post.R
\name{print.trim}
\alias{print.trim}
\title{print a 'trim' object}
\usage{
\method{print}{trim}(x, ...)
}
\arguments{
\item{x}{a \code{\link{trim}} object}

\item{...}{currently unused}
}
\description{
print a 'trim' object
}
\keyword{internal}
