% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_wald.R
\name{print.trim.wald}
\alias{print.trim.wald}
\title{Print an object of class trim.wald}
\usage{
\method{print}{trim.wald}(x, ...)
}
\arguments{
\item{x}{An object of class \code{trim.wald}}
}
\description{
Print an object of class trim.wald
}
\keyword{internal}

