% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tcf.R
\name{print.trimcommand}
\alias{print.trimcommand}
\title{print a trimcommand object}
\usage{
\method{print}{trimcommand}(x, ...)
}
\arguments{
\item{x}{an R object}

\item{...}{optional parameters (ignored)}
}
\description{
print a trimcommand object
}
\keyword{internal}

