% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_outfile.R
\name{read_tof}
\alias{read_tof}
\title{Read a TRIM3 output file}
\usage{
read_tof(file)
}
\arguments{
\item{file}{\code{[character]} filename}
}
\value{
A character string of class \code{tof}
}
\description{
Read a TRIM3 output file
}
\seealso{
Other parse_output: 
\code{\link{get_coef}()},
\code{\link{get_estimation_method}()},
\code{\link{get_gof}()},
\code{\link{get_model}()},
\code{\link{get_n_site}()},
\code{\link{get_n_time}()},
\code{\link{get_overal_imputed_slope}()},
\code{\link{get_time_indices}()},
\code{\link{get_time_totals}()},
\code{\link{get_version}()},
\code{\link{get_wald}()},
\code{\link{read_vcv}()}
}
\concept{parse_output}
\keyword{internal}
