% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtrim-pkg.R
\docType{package}
\name{rtrim-package}
\alias{rtrim-package}
\title{Trend and Indices for Monitoring Data}
\description{
The TRIM model is used to estimate species populations based on frequent
(annual) counts at a varying collection of sites. The model is able to take
account of missing data by imputing prior to estimation of population totals.
The current package is a complete re-implementation of the \code{Delphi}
based
\href{https://www.cbs.nl/en-gb/society/nature-and-environment/indices-and-trends--trim--}{TRIM}
software developed at Statistics Netherlands by Jeroen Pannekoek.
}
\section{Getting started}{


Users of the original TRIM software can get started by following the
\href{../doc/rtrim_for_TRIM_users.html}{rtrim for trim users} or
following the \href{../doc/Skylark_example.html}{TRIM by example} manuals.
}

