% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_post.R
\name{serial_correlation}
\alias{serial_correlation}
\title{Extract serial correlation from TRIM object}
\usage{
serial_correlation(x)
}
\arguments{
\item{x}{An object of class \code{\link{trim}}}
}
\value{
The serial correlation coefficient if computed, otherwise \code{NULL}.
}
\description{
Extract serial correlation from TRIM object
}
\seealso{
Other analyses: \code{\link{coef.trim}},
  \code{\link{confint.trim}}, \code{\link{gof}},
  \code{\link{index}}, \code{\link{now_what}},
  \code{\link{overall}}, \code{\link{overdispersion}},
  \code{\link{plot.trim.index}},
  \code{\link{plot.trim.overall}}, \code{\link{results}},
  \code{\link{summary.trim}}, \code{\link{totals}},
  \code{\link{trim}}, \code{\link{vcov.trim}},
  \code{\link{wald}}
}
