% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tcf.R
\name{summary.trimbatch}
\alias{summary.trimbatch}
\title{summarize a trimbatch object}
\usage{
\method{summary}{trimbatch}(object, ...)
}
\arguments{
\item{object}{a trimbatch object}

\item{...}{options (ignored)}
}
\description{
summarize a trimbatch object
}
\keyword{internal}
