% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_refine.R
\name{trim_refine}
\alias{trim_refine}
\title{TRIM stepwise refinement}
\usage{
trim_refine(count, site.id, time.id, covars = list(), model = 2,
  serialcor = FALSE, overdisp = FALSE, changepoints = integer(0),
  weights = numeric(0))
}
\arguments{
\item{count}{a numerical vector of count data.}

\item{site.id}{a numerical vector time points for each count data point.}

\item{time.id}{a numerical vector time points for each count data point.}

\item{covars}{an optional list of covariates}

\item{model}{a model type selector}

\item{serialcor}{a flag indication of autocorrelation has to be taken into account.}

\item{overdisp}{a flag indicating of overdispersion has to be taken into account.}

\item{changepoints}{a numerical vector change points (only for Model 2)}
}
\value{
a list of class \code{trim}, that contains all output, statistiscs, etc.
  Usually this information is retrieved by a set of postprocessing functions
}
\description{
TRIM stepwise refinement
}
\keyword{internal}

