% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tcf.R
\name{read_tcf}
\alias{read_tcf}
\title{Read a TRIM command file}
\usage{
read_tcf(file)
}
\arguments{
\item{file}{Location of tcf file.}
}
\value{
An object of class \code{\link{TRIMcommand}}
}
\description{
Read a TRIM command file
}

