% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_refine.R
\name{trim_refine}
\alias{trim_refine}
\title{TRIM stepwise refinement}
\usage{
trim_refine(count, site, year, month, weights, covars, model, changepoints,
  ..., premove = 0.2, penter = 0.15)
}
\arguments{
\item{count}{a numerical vector of count data.}

\item{site}{a vector (numerical or factor) of site identifiers for each count data point.}

\item{year}{a numerical vector of annual time points for each count data point.}

\item{month}{an optional numerical vector of monthly time points.}

\item{weights}{an optional numerical vector of weights.}

\item{covars}{an optional list of covariates.}

\item{model}{a model type selector.}

\item{changepoints}{a numerical vector change points (only for Model 2)}

\item{premove}{threshold probability for removal of parameters.}

\item{penter}{threshold probability for re-introduction of parameters.}
}
\value{
a list of class \code{trim}, that opcontains all output, statistiscs, etc.
  Usually this information is retrieved by a set of postprocessing functions
}
\description{
TRIM stepwise refinement
}
\keyword{internal}
